% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.r
\name{new_wmata}
\alias{new_wmata}
\title{Create a new wmata object}
\usage{
new_wmata(json, path)
}
\arguments{
\item{json}{list data from json api}

\item{path}{path to location of data}
}
\description{
Create a new wmata object
}
\keyword{interndal}
