% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-rail-station-info.r
\name{rail-station-information}
\alias{rail-station-information}
\alias{station_list}
\alias{station_timings}
\title{Rail Station Information}
\usage{
station_list(line_code = NULL)

station_timings(station_code = NULL)
}
\arguments{
\item{line_code}{Two-letter line code abbreviation (optional)}

\item{station_code}{Station code. Use the Station List method to return a list of all station codes. (optional)}
}
\description{
Rail line and station information, including locations, fares, times, and parking.
}
\section{Functions}{
\itemize{
\item \code{station_list}: Returns a list of station location and address information based on a given LineCode.
Omit the LineCode to return all stations. The response is an array of objects
identical to those returned in the Station Information method.

\item \code{station_timings}: Returns opening and scheduled first/last
train times based on a given StationCode. Omit the StationCode to return
timing information for all stations. Note that for stations with multiple
platforms (e.g.: Metro Center, L'Enfant Plaza, etc.), a distinct call is
required for each StationCode to retrieve the full set of train times at
such stations.
}}

\references{
\url{https://developer.wmata.com/docs/}
}
