% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-real-time-rail-predictions.r
\name{real-time-rail-predictions}
\alias{real-time-rail-predictions}
\alias{next_trains}
\title{Real-Time Rail Predictions}
\usage{
next_trains(station_codes)
}
\arguments{
\item{station_codes}{Character vector of station codes. For all predictions, use "All".}
}
\description{
Real-time rail prediction methods.
}
\section{Functions}{
\itemize{
\item \code{next_trains}: Returns next train arrival information for one or more stations. Will return an empty set of results when no predictions are available. Use All for the StationCodes parameter to return predictions for all stations.
}}

\references{
\url{https://developer.wmata.com/docs/}
}
