% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wmata-package.R
\docType{data}
\name{stations}
\alias{stations}
\title{Station location and address information}
\format{A data frame with 95 rows and 11 variables:}
\source{
\url{https://developer.wmata.com/}
}
\usage{
stations
}
\description{
A dataset containing the name, code, and other attributes of WMATA rail stations.
}
\keyword{datasets}
