% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wmata-package.R
\docType{package}
\name{wmata-package}
\alias{wmata}
\alias{wmata-package}
\title{wmata: WMATA API}
\description{
Provides an R API to the WMATA API.
}
\author{
\strong{Maintainer}: Robert Zamora \email{zamora.rr@gmail.com}

}
\keyword{internal}
